requirements.txt
# Requirements for JPEG text extraction and Azure OpenAI
Pillow>=9.0.0
pytesseract>=0.3.10
requests>=2.28.0
python-dotenv>=1.0.0

# Google Calendar API integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

annotated-types==0.7.0
anyio==3.7.1
asttokens==3.0.0
azure-core==1.36.0
azure-identity==1.25.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
executing==2.2.1
google-api-core==2.27.0
googleapis-common-protos==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
msal==1.34.0
msal-extensions==1.3.1
nest-asyncio==1.6.0
oauthlib==3.3.1
openai==1.3.0
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14